+++ packages/preview-middleware-client/test/unit/adp/quick-actions/fe-v4.test.ts
@@ -1047,7 +1047,7 @@ describe('FE V4 quick actions', () => {
                                     enabled: true,
                                     kind: 'simple',
                                     id: 'listReport0-enable-semantic-date-range',
-                                    title: 'Disable "Semantic Date Range" Button in Filter Bar'
+                                    title: 'Disable Semantic Date Range in Filter Bar'
                                 }
                             ]
                         }
@@ -1106,7 +1106,7 @@ describe('FE V4 quick actions', () => {
                                     enabled: true,
                                     kind: 'simple',
                                     id: 'listReport0-enable-semantic-date-range',
-                                    title: 'Enable "Semantic Date Range" for Filter Bar'
+                                    title: 'Enable Semantic Date Range in Filter Bar'
                                 }
                             ]
                         }