+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.3.17
+
+### Patch Changes
+
+-   Updated dependencies [cfdd442]
+    -   @sap-ux/project-input-validator@0.3.4
+
 ## 0.3.16
 
 ### Patch Changes